backend/requirements.txt
Flask
Flask-Cors
Flask-PyMongo
bson
python-dotenv
